@shirok 2024-07-12
これは困りものですね。多分大抵のアプリケーションではparallel mark使いたいと思うんですが…
オプションで切り替えるのは多分煩雑で、ユーザが気にせずともいい感じに動いて欲しいですね。
ちょっといいバランスを調査してみます。