@shirok 2025-09-10
久々に調べてみたんですが、ベンチマークのタイムアウトは実時間じゃなくてCPU時間見てるんですね。CPU時間は複数スレッド並列で走らせればそれが全部加算されますから、parallel mark使うと実時間以上に増えます。

Linux上では、実時間で比べると0.9.14でcpstakが111sかかっていたのが0.9.15以降だと65s前後になり、体感上の速度はparallel markの方が明らかに向上してます。しかしCPU時間(user+sys)は0.9.14では実時間とほぼ同じなのが、parallel markを使うと327sになり、benchスクリプトの制限300sを超えるのでタイムアウトになります。

Windowsでの計測でどちらを測っていたかわかりませんが、実用上は実時間が短くなる方を選択すべきだと思います。

ベンチマークの方で、並列性によるCPU時間増加を織り込んでくれるとありがたいんですが、Gaucheの最適化を進めることで全体の実行時間を下げて制限をクリアする方が建設的かもしれません。