@shirok 2025-09-16
Windowsではデフォルトでmark threadsを初期化時にspawnしないようにしてみました。環境変数`GAUCHE_PARALLEL_MARK_ALWAYS` が定義されていたらUnixと同じ動作になります。
(なお、`thread-start!`でユーザコードからスレッドを作った場合はmark threadもその時作られます。)